v0.12.2: pay down lint debt that accumulated across the _sre and dict-binding work
The lint job on this branch has been red since the dict-binding and _sre
landings: golangci-lint v2.12.2 flagged unparam, nilerr, gocognit, and a
clutch of gocritic/staticcheck nits, none of which Phase 1 introduced
but all of which gate the green CI run that the Phase 1 comment needs.

Mechanical fixes:
- drop unused error returns from dictHasKeys, setUnion, and lookupIn,
  fold their call sites accordingly
- mark the two intentional miss-as-False returns with //nolint:nilerr
  (function.__isabstractmethod__ getter, lookupIn mapping fallback)
- drop the redundant uint32 conversions in vm/eval_gen.go
- rewrite the three if/else-if chains as switch statements (codegen,
  str_bind.expandtabs, usertype attr-slot pick)
- collapse []rune(s) + range to range s in str_bind.translate
- tagged switch on the rune in str_bind.expandtabs

Decomposition fixes:
- split registerFunctionGetSets into identity / read-only / typeparams /
  dict groups, plus pull __isabstractmethod__ and __dict__ getters and
  setters into top-level funcs
- split strMakeTrans into makeTransFromDict and makeTransFromStrings
- pull typeCall's TpNew path and abstract-class check into helpers
- split fixupSlotDispatchers into one helper per slot group (call/repr,
  inherit-from-base, hash/iter, richcmp/bool, subscript, descriptor)

+++ vm/eval_gen.go
@@ -167,11 +167,11 @@ func (e *evalState) execReturnGenerator() (genResult, error) {
 		retVal  objects.Object
 	)
 	switch {
-	case flags&uint32(compile.CoCoroutine) != 0:
+	case flags&compile.CoCoroutine != 0:
 		c := objects.NewCoroutine(name)
 		yieldCh, sendCh = c.YieldCh, c.SendCh
 		retVal = c
-	case flags&uint32(compile.CoAsyncGenerator) != 0:
+	case flags&compile.CoAsyncGenerator != 0:
 		ag := objects.NewAsyncGenerator(name)
 		yieldCh, sendCh = ag.YieldCh, ag.SendCh
 		retVal = ag
